Fix for new position of tabs caused by folders
Added option to set a tab as "in a folder" because the tab is lower on the screen
Only show tabs that are supported (at the moment Normal and Premium)
Change font color of tab to be readable no matter the background color.
Add some hardcoded gem colors like Splitting Steel where it's a Dex gem but the Str requirement is higher.
Made settings save by tab ID rather than Index so that if you rearrange your tabs it will keep the settings where they should be.